+++ packages/jaspr_test/lib/browser_test.dart
@@ -1,2 +1,3 @@
 export 'jaspr_test.dart';
+export 'src/browser_matchers.dart';
 export 'src/testers/browser_tester.dart';