+++ packages/jaspr_test/lib/src/testers/fake_event_web.dart
@@ -1,0 +1,5 @@
+import 'package:web/web.dart';
+
+Event fakeEvent() {
+  return Event('click');
+}