+++ packages/jaspr/lib/src/foundation/events/events_vm.dart
@@ -1,4 +1,0 @@
-typedef Event = dynamic;
-typedef InputElement = dynamic;
-typedef TextAreaElement = dynamic;
-typedef SelectElement = dynamic;