+++ examples/flutter_embedding/lib/components/flutter_app_container.dart
@@ -1,5 +1,10 @@
 import 'package:jaspr/jaspr.dart';
+import 'package:jaspr_riverpod/jaspr_riverpod.dart';
 
+@Import.onWeb('package:flutter_riverpod/flutter_riverpod.dart', show: [#UncontrolledProviderScope])
+@Import.onWeb('../widgets/app.dart', show: [#MyApp])
+import 'flutter_app_container.imports.dart' as flt;
+import 'flutter_target.dart';
 import 'ripple_loader.dart';
 
 class FlutterAppContainer extends StatelessComponent {
@@ -7,10 +12,14 @@ class FlutterAppContainer extends StatelessComponent {
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield article([
-      div(id: 'flutter_target', [
-        RippleLoader(),
-      ])
-    ]);
+    yield FlutterTarget(
+      loader: RippleLoader(),
+      app: kIsWeb
+          ? flt.UncontrolledProviderScope(
+              container: ProviderScope.containerOf(context),
+              child: flt.MyApp(),
+            )
+          : null,
+    );
   }
 }