+++ examples/flutter_multi_view/lib/components/embedded_counter.imports.dart
@@ -1,0 +1,5 @@
+// GENERATED FILE, DO NOT MODIFY
+// Generated with jaspr_builder
+
+export '../generated/imports/_web.dart' if (dart.library.io) '../generated/imports/_stubs.dart'
+    show CounterWidget, CounterWidgetOrStubbed;