+++ packages/jaspr/lib/src/foundation/view_transition/view_transition_io.dart
@@ -1,0 +1,5 @@
+import 'dart:async';
+
+FutureOr<void> startViewTransition(void Function() callback) {
+  callback();
+}