+++ packages/jaspr_flutter_embed/lib/src/run_flutter_app.dart
@@ -2,27 +2,48 @@
 @JS()
 library flutter_interop;
 
+import 'dart:async';
 import 'dart:js_interop';
 import 'dart:ui_web' as ui_web;
 
+import 'package:flutter/widgets.dart' as flt;
 import 'package:web/web.dart';
 
-/// Starts a flutter app and attaches it to the [attachTo] dom element.
-void runFlutterApp({required String attachTo, required void Function() runApp}) {
-  var target = document.querySelector(attachTo);
+import 'multi_view_app.dart';
+import 'view_constraints.dart';
 
-  flutter ??= FlutterInterop();
-  flutter!.loader ??= FlutterLoader();
+Map<int, flt.Widget> _viewWidgets = {};
 
-  flutter!.loader!.didCreateEngineInitializer = ((engineInitializer) {
-    return engineInitializer
-        .initializeEngine(InitializeEngineOptions(hostElement: target!))
-        .toDart
-        .then((runner) => runner.runApp())
-        .toJS;
+Future<FlutterApp> _flutterApp = Future(() {
+  final completer = Completer<FlutterApp>();
+
+  flutter!.loader!.didCreateEngineInitializer = (EngineInitializer engineInitializer) {
+    return Future(() async {
+      var engine = await engineInitializer.initializeEngine(InitializeEngineOptions(multiViewEnabled: true)).toDart;
+      var app = await engine.runApp().toDart;
+      completer.complete(app);
+    }).toJS;
+  }.toJS;
+
+  ui_web.bootstrapEngine(runApp: () {
+    flt.runWidget(MultiViewApp(viewBuilder: (viewId) => _viewWidgets[viewId]));
   });
 
-  ui_web.bootstrapEngine(runApp: runApp);
+  return completer.future;
+});
+
+Future<void> preloadEngine() => _flutterApp;
+
+Future<int> addView(Element target, ViewConstraints? constraints, flt.Widget widget) async {
+  var app = await _flutterApp;
+  var id = app.addView(ViewOptions(hostElement: target, viewConstraints: constraints));
+  _viewWidgets[id] = widget;
+  return id;
+}
+
+Future<void> removeView(int viewId) async {
+  var app = await _flutterApp;
+  app.removeView(viewId);
 }
 
 /// Handle to the [_flutter] object defined by the 'flutter.js' script.
@@ -35,31 +56,49 @@ external FlutterInterop? flutter;
 extension type FlutterInterop._(JSObject _) {
   external FlutterInterop({FlutterLoader? loader});
 
-  external FlutterLoader? get loader;
-  external set loader(FlutterLoader? loader);
+  external FlutterLoader? loader;
 }
 
 extension type FlutterLoader._(JSObject _) {
   external FlutterLoader({JSFunction? didCreateEngineInitializer});
 
-  @JS("didCreateEngineInitializer")
-  external set _didCreateEngineInitializer(JSFunction? fn);
+  external JSPromise<JSAny?> load(LoadOptions options);
 
-  set didCreateEngineInitializer(OnEntrypointLoaded? callback) {
-    _didCreateEngineInitializer = callback?.toJS;
-  }
+  external JSFunction? didCreateEngineInitializer;
 }
 
-typedef OnEntrypointLoaded = JSPromise Function(EngineInitializer engineInitializer);
+extension type LoadOptions._(JSObject _) {
+  external LoadOptions({
+    String? entrypointUrl,
+    JSFunction? onEntrypointLoaded,
+  });
+}
 
 extension type EngineInitializer._(JSObject _) {
+  external EngineInitializer();
+
   external JSPromise<AppRunner> initializeEngine(InitializeEngineOptions options);
 }
 
 extension type InitializeEngineOptions._(JSObject _) {
-  external InitializeEngineOptions({Element? hostElement});
+  external InitializeEngineOptions({
+    Element? hostElement,
+    bool? multiViewEnabled,
+  });
 }
 
 extension type AppRunner._(JSObject _) implements JSObject {
-  external JSPromise<JSAny> runApp();
+  external AppRunner();
+
+  external JSPromise<FlutterApp> runApp();
+}
+
+extension type ViewOptions._(JSObject _) {
+  external ViewOptions({Element? hostElement, JSAny? initialData, ViewConstraints? viewConstraints});
+}
+
+extension type FlutterApp._(JSObject _) implements JSObject {
+  external int addView(ViewOptions options);
+
+  external void removeView(int viewId);
 }