+++ packages/jaspr_flutter_embed/lib/src/widget/widget_io.dart
@@ -1,0 +1,1 @@
+typedef Widget = Never;