+++ packages/jaspr_flutter_embed/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased breaking
+
+- **BREAKING** Migrate to `package:web`.
+
 ## 0.3.4
 
 - `jaspr` upgraded to `0.15.0`